Implement obstacle placement (vimkey layout)
Implement a method of placing obstacles adjacent to the players current
position. Obstacles can only placed if Player.obstacle>0 and once it is
placed, this attribute is decremented.
The controls for this mechanic are mapped in a "vim key"-layout:

H = Left
J = Down
K = Up
L = Right

Furthermore, the KeyMapper needed to be adjusted to allow this behavior.
It now registers Event* instead of Event instances.
The Tilemaps have also for this purpose been fitted with the following method:

    get_coordinate_from_offset(float x, float y